code/backend/requirements.txt
asgiref==3.6.0
cmake==3.25.2
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
mysql-client==0.0.1
mysqlclient==2.1.1
numpy==1.24.2
openpyxl==3.1.2
pandas==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
